fix main-menu file-load-error scrolling bug
This bug occurs when a file fails to load.
When scrolling up or down, the messages for files failing
to load stay in their original places even when they should not.
This is because the error grid is not moved along with the menu
grid upon scrolling. This is fixed by moving the error grid.